requirements.txt
numpy
pandas
setuptools
google-cloud-storage
pyyaml
ipykernel
matplotlib
wordcloud 
scikit-learn
comet-ml
dvc
tf-nightly-cpu